// src/types/generated/DecisionQAPair.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { QAPairContext } from "./QAPairContext";

export interface DecisionQAPair { qaIndex: number, assistantAnswerUuid: string, userDecisionUuid: string, assistantAnswer: string, userDecision: string, contextQaPairs?: Array<QAPairContext>, }