// src/types/generated/DecisionAnalysis.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Alternative } from "./Alternative";
import type { DecisionType } from "./DecisionType";

export interface DecisionAnalysis { decisionMade: string, decisionType: DecisionType, techStack: Array<string>, rationale: Array<string>, inferredReasons: Array<string>, alternatives: Array<Alternative>, confidence: number, }